Refactor feed data and content field usage
Removes feed-related actions and selectors from the store and hooks, updates field exports to use a static contentField instead of a factory, and standardizes import paths and naming. This simplifies data management to only following and interactions, and streamlines field usage in feed views.
