+++ src/social-web/store/actions.ts
@@ -6,20 +6,13 @@ import apiFetch from '@wordpress/api-fetch';
 /**
  * Internal dependencies
  */
-import type { Following, Interaction, FeedPost } from '../types';
-import type { SetFollowingAction, SetInteractionsAction, SetFeedAction, SetLoadingAction, State } from './types';
+import type { Following, Interaction } from '../types';
+import type { SetFollowingAction, SetInteractionsAction, SetLoadingAction, State } from './types';
 
 /**
  * Store actions
  */
 export const actions = {
-	setFeed( feed: FeedPost[] ): SetFeedAction {
-		return {
-			type: 'SET_FEED',
-			feed,
-		};
-	},
-
 	setFollowing( following: Following[] ): SetFollowingAction {
 		return {
 			type: 'SET_FOLLOWING',
@@ -67,22 +60,6 @@ export const actions = {
 			console.error( 'Failed to fetch interactions:', error );
 		} finally {
 			yield actions.setLoading( 'interactions', false );
-		}
-	},
-
-	*fetchFeed() {
-		yield actions.setLoading( 'feed', true );
-		try {
-			const feed = yield apiFetch( {
-				path: '/wp/v2/ap_post',
-			} );
-			console.log( 'fetchFeed: Received data:', feed );
-			yield actions.setFeed( Array.isArray( feed ) ? feed : [] );
-		} catch ( error ) {
-			console.error( 'Failed to fetch feed:', error );
-			yield actions.setFeed( [] );
-		} finally {
-			yield actions.setLoading( 'feed', false );
 		}
 	},
 };
