+++ src/social-web/store/reducer.ts
@@ -9,12 +9,6 @@ import { DEFAULT_STATE } from './types';
  */
 export function reducer( state = DEFAULT_STATE, action: Action ): State {
 	switch ( action.type ) {
-		case 'SET_FEED':
-			return {
-				...state,
-				feed: action.feed,
-			};
-
 		case 'SET_FOLLOWING':
 			return {
 				...state,
